// client/src/store/auth/actionTypes.ts
export enum ActionTypes {
  AUTH_START = "AUTH:AUTH_START",
  AUTH_FAIL = "AUTH:AUTH_FAIL",
  REGISTRATION_SUCCESS = "AUTH:REGISTRATION_SUCCESS",
  LOGIN_SUCCESS = "AUTH:LOGIN_SUCCESS",
  LOGOUT = "AUTH:LOGOUT",
  FETCH_USER_START = "AUTH:FETCH_USER_START",
  FETCH_USER_SUCCESS = "AUTH:FETCH_USER_SUCCESS",
  FETCH_USER_FAIL = "AUTH:FETCH_USER_FAIL",
  CHANGE_PASSWORD_START = "AUTH:CHANGE_PASSWORD_START",
  CHANGE_PASSWORD_FAIL = "AUTH:CHANGE_PASSWORD_FAIL",
  CHANGE_PASSWORD_SUCCESS = "AUTH:CHANGE_PASSWORD_SUCCESS",
  SET_USER_ONLINE = "AUTH:SET_USER_ONLINE",
}